Move custom colour scheme class to color_scheme.dart file
