+++ app/lib/common/widgets/spaces/space_info.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/providers/space_providers.dart';
-import 'package:acter/common/themes/app_theme.dart';
+import 'package:acter/common/themes/colors/color_scheme.dart';
 import 'package:acter/common/utils/utils.dart';
 import 'package:acter/common/widgets/visibility/visibility_chip.dart';
 import 'package:acter_flutter_sdk/acter_flutter_sdk_ffi.dart';
