+++ app/lib/features/onboarding/pages/forgot_password.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/providers/sdk_provider.dart';
-import 'package:acter/common/themes/app_theme.dart';
+import 'package:acter/common/themes/colors/color_scheme.dart';
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/utils/routes.dart';
 import 'package:acter/common/utils/validation_utils.dart';
