+++ app/lib/features/onboarding/pages/register_page.dart
@@ -1,5 +1,5 @@
 import 'package:acter/common/providers/network_provider.dart';
-import 'package:acter/common/themes/app_theme.dart';
+import 'package:acter/common/themes/colors/color_scheme.dart';
 import 'package:acter/common/toolkit/buttons/inline_text_button.dart';
 
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
